+++ bdsim.cc
@@ -210,13 +210,11 @@ int main(int argc,char** argv)
 #ifdef BDSDEBUG 
   G4cout << __FUNCTION__ << "> Registering user action - Stepping Action"<<G4endl;
 #endif
-  // Only add steppingaction if it is actually used, so do check here (for cpu reasons)
-  G4bool verboseStepping = globalConstants->VerboseStep();
-  if (globalConstants->ThresholdCutPhotons() > 0 || globalConstants->ThresholdCutCharged() > 0
-      || verboseStepping)
+  // Only add steppingaction if it is actually used, so do check here (for performance reasons)
+  if (globalConstants->VerboseStep())
     {
       G4int verboseEventNumber = globalConstants->VerboseEventNumber();
-      runManager->SetUserAction(new BDSSteppingAction(verboseStepping, verboseEventNumber));
+      runManager->SetUserAction(new BDSSteppingAction(true, verboseEventNumber));
     }
   
 #ifdef BDSDEBUG