+++ parser/optionsBase.cc
@@ -158,8 +158,6 @@ OptionsBase::OptionsBase()
   maximumPhotonsPerStep    = -1;  ///< -1 -> no action taken (could want 0)
   maximumBetaChangePerStep = 10;
   maximumTracksPerEvent    = 0;   ///< 0 -> no action taken
-  thresholdCutCharged      = 0.0;
-  thresholdCutPhotons      = 0.0;
   defaultRangeCut          = 1e-3;
   prodCutPhotons           = 1e-3;
   prodCutElectrons         = 1e-3;